requirements.txt
Flask==3.0.3
numpy==1.24.3
opencv-python-headless==4.5.5.64
tensorflow-cpu==2.13.1
joblib==1.2.0
boto3==1.24.23
python-dotenv==0.20.0
